requirements.txt
python-telegram-bot==20.3
gspread
oauth2client
python-dotenv
telegram
python-telegram-bot
python-telegram-bot==13.0
python-telegram-bot
